make component async
