+++ .babelrc
@@ -2,14 +2,14 @@
     "presets": ["@babel/preset-env", "@babel/preset-react"],
     "env": {
         "production": {
-            "plugins": ["@babel/plugin-proposal-object-rest-spread", "styled-jsx/babel"]
+            "plugins": ["@babel/plugin-proposal-object-rest-spread", "styled-jsx/babel", "@babel/plugin-syntax-dynamic-import"]
         },
         "development": {
-            "plugins": ["@babel/plugin-proposal-object-rest-spread", "styled-jsx/babel"]
+            "plugins": ["@babel/plugin-proposal-object-rest-spread", "styled-jsx/babel", "@babel/plugin-syntax-dynamic-import"]
         },
         "test": {
-            "plugins": ["@babel/plugin-proposal-object-rest-spread", "styled-jsx/babel-test"]
+            "plugins": ["@babel/plugin-proposal-object-rest-spread", "styled-jsx/babel-test", "@babel/plugin-syntax-dynamic-import"]
         }
     }
   }
-  
+
