+++ package.json
@@ -30,8 +30,10 @@
   "devDependencies": {
     "@babel/core": "^7.19.1",
     "@babel/plugin-proposal-object-rest-spread": "^7.18.9",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/preset-env": "^7.19.1",
     "@babel/preset-react": "^7.18.6",
+    "@plotly/webpack-dash-dynamic-import": "^1.3.0",
     "@types/node": "^18.7.18",
     "babel-eslint": "^10.0.2",
     "babel-loader": "^8.2.5",
